+++ src/main/java/uk/gov/hmcts/darts/courthouse/api/CourthouseApi.java
@@ -1,8 +1,8 @@
-package uk.gov.hmcts.darts.courthouses.api;
+package uk.gov.hmcts.darts.courthouse.api;
 
 import uk.gov.hmcts.darts.common.entity.CourthouseEntity;
-import uk.gov.hmcts.darts.courthouses.exception.CourthouseCodeNotMatchException;
-import uk.gov.hmcts.darts.courthouses.exception.CourthouseNameNotFoundException;
+import uk.gov.hmcts.darts.courthouse.exception.CourthouseCodeNotMatchException;
+import uk.gov.hmcts.darts.courthouse.exception.CourthouseNameNotFoundException;
 
 public interface CourthouseApi {
     CourthouseEntity retrieveAndUpdateCourtHouse(Integer courthouseCode, String courthouseName)
