+++ src/main/java/uk/gov/hmcts/darts/courthouse/api/impl/CourthouseApiImpl.java
@@ -1,12 +1,12 @@
-package uk.gov.hmcts.darts.courthouses.api.impl;
+package uk.gov.hmcts.darts.courthouse.api.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.darts.common.entity.CourthouseEntity;
-import uk.gov.hmcts.darts.courthouses.api.CourthouseApi;
-import uk.gov.hmcts.darts.courthouses.exception.CourthouseCodeNotMatchException;
-import uk.gov.hmcts.darts.courthouses.exception.CourthouseNameNotFoundException;
-import uk.gov.hmcts.darts.courthouses.service.CourthouseService;
+import uk.gov.hmcts.darts.courthouse.api.CourthouseApi;
+import uk.gov.hmcts.darts.courthouse.exception.CourthouseCodeNotMatchException;
+import uk.gov.hmcts.darts.courthouse.exception.CourthouseNameNotFoundException;
+import uk.gov.hmcts.darts.courthouse.service.CourthouseService;
 
 @RequiredArgsConstructor
 @Service
