+++ src/main/java/uk/gov/hmcts/darts/courthouse/controller/CourthousesController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.courthouses.controller;
+package uk.gov.hmcts.darts.courthouse.controller;
 
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -12,10 +12,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.darts.common.entity.CourthouseEntity;
 import uk.gov.hmcts.darts.courthouse.api.CourthousesApi;
+import uk.gov.hmcts.darts.courthouse.mapper.CourthouseToCourthouseEntityMapper;
 import uk.gov.hmcts.darts.courthouse.model.Courthouse;
 import uk.gov.hmcts.darts.courthouse.model.ExtendedCourthouse;
-import uk.gov.hmcts.darts.courthouses.mapper.CourthouseToCourthouseEntityMapper;
-import uk.gov.hmcts.darts.courthouses.service.CourthouseService;
+import uk.gov.hmcts.darts.courthouse.service.CourthouseService;
 
 import java.util.List;
 import javax.validation.Valid;
