+++ src/main/java/uk/gov/hmcts/darts/courthouse/exception/CourthouseCodeNotMatchException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.courthouses.exception;
+package uk.gov.hmcts.darts.courthouse.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
