+++ src/main/java/uk/gov/hmcts/darts/courthouse/service/CourthouseServiceImpl.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.courthouses.service;
+package uk.gov.hmcts.darts.courthouse.service;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -7,12 +7,12 @@ import org.springframework.stereotype.Service;
 import uk.gov.hmcts.darts.common.entity.CourthouseEntity;
 import uk.gov.hmcts.darts.common.exception.DartsApiException;
 import uk.gov.hmcts.darts.common.service.RetrieveCoreObjectService;
+import uk.gov.hmcts.darts.courthouse.CourthouseRepository;
+import uk.gov.hmcts.darts.courthouse.exception.CourthouseApiError;
+import uk.gov.hmcts.darts.courthouse.exception.CourthouseCodeNotMatchException;
+import uk.gov.hmcts.darts.courthouse.exception.CourthouseNameNotFoundException;
+import uk.gov.hmcts.darts.courthouse.mapper.CourthouseToCourthouseEntityMapper;
 import uk.gov.hmcts.darts.courthouse.model.Courthouse;
-import uk.gov.hmcts.darts.courthouses.CourthouseRepository;
-import uk.gov.hmcts.darts.courthouses.exception.CourthouseApiError;
-import uk.gov.hmcts.darts.courthouses.exception.CourthouseCodeNotMatchException;
-import uk.gov.hmcts.darts.courthouses.exception.CourthouseNameNotFoundException;
-import uk.gov.hmcts.darts.courthouses.mapper.CourthouseToCourthouseEntityMapper;
 
 import java.util.List;
 import java.util.Objects;
